Disallow all but the most current API/Lore documentation from being indexed by search engines
This change would disallow Google/Bing/et al from indexing old versions of the documentation, and just 'current'. The upside of this is that googling for "twisted api documentation" will then bring up the most recent always (and not 13.1.0 like it does currently). The downside to this is that it makes searching for older versions of the API documentation much harder (eg. "twisted 8.2 api documentation" wouldn't come up with anything).